tests/submissions/anssik/battery-full.html

<!DOCTYPE html>
<html>
  <head>
  <title>Battery Status API Test Suite</title>
  <script src="http://www.w3c-test.org/resources/testharness.js"></script>
  <link rel='stylesheet' href='http://w3c-test.org/resources/testharness.css' media='all'/>
  </head>
  <body>
    <h1>Description</h1>
    <p>
      This test validates that all of the navigator.battery attributes exist and are set to correct values, when the battery is full.
    </p>
    <h2>Preconditions</h2>
    <ol>
      <li>
        The device is plugged in to the charger before this test case is run.
      </li>
      <li>
        The battery is full.
      </li>
    </ol>
    <div id="log"></div>
    <script>
    (function() {

      test(function() {
        assert_true(navigator.battery.charging);
      }, 'The charging attribute must be set to true if the battery is charging ("or otherwise" condition).');
      
      test(function() {
        assert_true(navigator.battery.chargingTime === 0);
      }, 'The chargingTime attribute must be set to 0, if the battery is full.');
      
      test(function() {
        assert_true(navigator.battery.dischargingTime === Infinity);
      }, 'The dischargingTime attribute must be set to the value positive Infinity if the battery is charging ("or otherwise" condition).');
      
      test(function() {
        assert_true(navigator.battery.level === 1.0);
      }, 'The level attribute must be set to the current battery level scaled from 0 to 1.0.');
      
    })();
    </script>
  </body>
</html>
